refactor: ViewModelを作り、通信とViewのロジックを分離。
